Struct mousetrap::mousetrap::MouseArena

pub struct MouseArena {
    // some fields omitted
}

A MouseArena is the area where the mouse can still be moved to.

Allows to warp the mouse to the centre of the area (trap it) or further reduce it's size (close in on it).

Methods

impl MouseArena

fn new(width: u16, height: u16, h_min: u8, v_min: u8) -> MouseArena

Create a new MouseArena spanning the entire screen.

fn close_in(&mut self, dir: TrapDirection) -> bool

Reduce the MouseArea by a factor of two, given a direction.

fn trap(&self) -> (u16, u16)

Warp the mouse to the centre of the current MouseArena.